separate codegen and message desc spec
So far we used the type naming scheme
of the Cast gem both for the code
that mangles the result of Cast parsing
and for indicating components in the
message description.

While this is an obvious choice when
processing Cast output, there is no
logical need for this for the message
description. So we introduce a more
intuitive message component naming
scheme:
- use "buf" instead of "char"
- use false to indicate having a special
  handler for given message.

Also improved the comments in the message
description.

Implied, the processing code accordingly
diverges for the two cases, so there
is no more branching on "[0]char" just
because that's the adequate Go type
for a buffer placeholder struct field.