ruby: add rudimentary support for v2 dumps
We don't implement parsing the metadata, just add
a --meta-size option that can be used to inform the
program to consider a metadata block of that size for
each message. Obviously it works only if the metadata
size is fix throughout the dump, and the user has to
figure out this size by herself prior to using the
program.