pre-format timestamp
Instead of placing time.Time objects into the message
structure and represent it in RFC3339 format only upon
human readable formatting, put the RFC3339 string into
the message right away.

This is a workaround to get time represented
as RFC3339 in msgpack output instead of binary blob.
If we serialized to formats with proper support for
temporal data, this approach would be incorrect.

As far as JSON is concerned, this change does not
make a difference, as JSON codec applies RFC3339
representation for temporal data anyway.

If we add serialization to a time aware format,
this change will have to undone and take the pain
to represent temporal data in a suitable way for
all the non-time-aware formatters.

(We experimented with CBOR, a contender of msgpack
with time support, but we found that the implementations
are less mature and in particular they don't ser/deser
temporal data properly. Maybe it will be worth to revisit
later.)